+++ programs/drift/src/controller/orders/tests.rs
@@ -124,7 +124,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -242,7 +241,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -360,7 +358,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -478,7 +475,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -596,7 +592,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -680,7 +675,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -765,7 +759,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -850,7 +843,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -935,7 +927,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -1040,7 +1031,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -1148,7 +1138,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -1263,7 +1252,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            0,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -1379,7 +1367,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -1519,7 +1506,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -1634,7 +1620,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -1733,7 +1718,6 @@ pub mod fulfill_order_with_maker_order {
             Some(oracle_price),
             now,
             slot,
-            10,
             &fee_structure,
             &mut oracle_map,
             &mut order_records,
@@ -1876,7 +1860,6 @@ pub mod fulfill_order_with_maker_order {
             Some(oracle_price),
             now,
             slot,
-            10,
             &fee_structure,
             &mut oracle_map,
             &mut order_records,
@@ -2007,7 +1990,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut oracle_map,
             &mut order_records,
@@ -2149,7 +2131,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut oracle_map,
             &mut order_records,
@@ -2190,204 +2171,6 @@ pub mod fulfill_order_with_maker_order {
         assert_eq!(market.amm.total_fee, 20000);
         assert_eq!(market.amm.total_fee_minus_distributions, 20000);
         assert_eq!(market.amm.net_revenue_since_last_funding, 20000);
-    }
-
-    #[test]
-    fn taker_limit_bid_fails_to_cross_because_of_vamm_guard() {
-        let now = 5_i64;
-        let slot = 5_u64;
-
-        let mut maker = User {
-            orders: get_orders(Order {
-                market_index: 0,
-                post_only: true,
-                order_type: OrderType::Limit,
-                direction: PositionDirection::Short,
-                base_asset_amount: BASE_PRECISION_U64,
-                slot: 0,
-                price: 150 * PRICE_PRECISION_U64,
-                ..Order::default()
-            }),
-            perp_positions: get_positions(PerpPosition {
-                market_index: 0,
-                open_orders: 1,
-                open_asks: -BASE_PRECISION_I64,
-                ..PerpPosition::default()
-            }),
-            ..User::default()
-        };
-
-        let mut taker = User {
-            orders: get_orders(Order {
-                market_index: 0,
-                order_type: OrderType::Limit,
-                direction: PositionDirection::Long,
-                base_asset_amount: BASE_PRECISION_U64,
-                price: 150 * PRICE_PRECISION_U64,
-                auction_duration: 10,
-                ..Order::default()
-            }),
-            perp_positions: get_positions(PerpPosition {
-                market_index: 0,
-                open_orders: 1,
-                open_bids: BASE_PRECISION_I64,
-                ..PerpPosition::default()
-            }),
-            ..User::default()
-        };
-
-        let mut oracle_price = get_pyth_price(100, 6);
-        let oracle_price_key =
-            Pubkey::from_str("J83w4HKfqxwcq3BEMMkPFSppX3gqekLyLJBexebFVkix").unwrap();
-        let pyth_program = crate::ids::pyth_program::id();
-        create_account_info!(
-            oracle_price,
-            &oracle_price_key,
-            &pyth_program,
-            oracle_account_info
-        );
-        let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
-
-        let mut market = PerpMarket::default_test();
-        market.amm.peg_multiplier = 100 * PEG_PRECISION;
-        market.amm.oracle = oracle_price_key;
-
-        let fee_structure = get_fee_structure();
-        let (maker_key, taker_key, filler_key) = get_user_keys();
-
-        let mut order_records = vec![];
-
-        let mut taker_stats = UserStats::default();
-        let mut maker_stats = UserStats::default();
-
-        let oracle_price = 100 * PRICE_PRECISION_I64;
-
-        let (base_asset_amount, _) = fulfill_perp_order_with_match(
-            &mut market,
-            &mut taker,
-            &mut taker_stats,
-            0,
-            &taker_key,
-            &mut maker,
-            &mut Some(&mut maker_stats),
-            0,
-            &maker_key,
-            &mut None,
-            &mut None,
-            &filler_key,
-            &mut None,
-            &mut None,
-            0,
-            Some(oracle_price),
-            now,
-            slot,
-            10,
-            &fee_structure,
-            &mut oracle_map,
-            &mut order_records,
-        )
-        .unwrap();
-
-        assert_eq!(base_asset_amount, 0);
-    }
-
-    #[test]
-    fn taker_limit_ask_fails_to_cross_because_of_vamm_guard() {
-        let now = 5_i64;
-        let slot = 5_u64;
-
-        let mut maker = User {
-            orders: get_orders(Order {
-                market_index: 0,
-                post_only: true,
-                order_type: OrderType::Limit,
-                direction: PositionDirection::Long,
-                base_asset_amount: BASE_PRECISION_U64,
-                slot: 0,
-                price: 50 * PRICE_PRECISION_U64,
-                ..Order::default()
-            }),
-            perp_positions: get_positions(PerpPosition {
-                market_index: 0,
-                open_orders: 1,
-                open_asks: -BASE_PRECISION_I64,
-                ..PerpPosition::default()
-            }),
-            ..User::default()
-        };
-
-        let mut taker = User {
-            orders: get_orders(Order {
-                market_index: 0,
-                order_type: OrderType::Limit,
-                direction: PositionDirection::Short,
-                base_asset_amount: BASE_PRECISION_U64,
-                price: 50 * PRICE_PRECISION_U64,
-                auction_duration: 10,
-                ..Order::default()
-            }),
-            perp_positions: get_positions(PerpPosition {
-                market_index: 0,
-                open_orders: 1,
-                open_bids: BASE_PRECISION_I64,
-                ..PerpPosition::default()
-            }),
-            ..User::default()
-        };
-
-        let mut oracle_price = get_pyth_price(100, 6);
-        let oracle_price_key =
-            Pubkey::from_str("J83w4HKfqxwcq3BEMMkPFSppX3gqekLyLJBexebFVkix").unwrap();
-        let pyth_program = crate::ids::pyth_program::id();
-        create_account_info!(
-            oracle_price,
-            &oracle_price_key,
-            &pyth_program,
-            oracle_account_info
-        );
-        let mut oracle_map = OracleMap::load_one(&oracle_account_info, slot, None).unwrap();
-
-        let mut market = PerpMarket::default_test();
-        market.amm.peg_multiplier = 100 * PEG_PRECISION;
-        market.amm.oracle = oracle_price_key;
-
-        let fee_structure = get_fee_structure();
-        let (maker_key, taker_key, filler_key) = get_user_keys();
-
-        let mut order_records = vec![];
-
-        let mut taker_stats = UserStats::default();
-        let mut maker_stats = UserStats::default();
-
-        let oracle_price = 100 * PRICE_PRECISION_I64;
-
-        let (base_asset_amount, _) = fulfill_perp_order_with_match(
-            &mut market,
-            &mut taker,
-            &mut taker_stats,
-            0,
-            &taker_key,
-            &mut maker,
-            &mut Some(&mut maker_stats),
-            0,
-            &maker_key,
-            &mut None,
-            &mut None,
-            &filler_key,
-            &mut None,
-            &mut None,
-            0,
-            Some(oracle_price),
-            now,
-            slot,
-            10,
-            &fee_structure,
-            &mut oracle_map,
-            &mut order_records,
-        )
-        .unwrap();
-
-        assert_eq!(base_asset_amount, 0);
     }
 
     #[test]
@@ -2473,7 +2256,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            10,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -2598,7 +2380,6 @@ pub mod fulfill_order_with_maker_order {
             None,
             now,
             slot,
-            0,
             &fee_structure,
             &mut get_oracle_map(),
             &mut order_records,
@@ -8509,7 +8290,6 @@ pub mod sanitize_maker_orders {
             &filler_key,
             0,
             oracle_price,
-            100 * PRICE_PRECISION_U64,
             clock.unix_timestamp,
             clock.slot,
         )
@@ -8709,7 +8489,6 @@ pub mod sanitize_maker_orders {
             &filler_key,
             0,
             oracle_price,
-            100 * PRICE_PRECISION_U64,
             clock.unix_timestamp,
             clock.slot,
         )
@@ -8898,7 +8677,6 @@ pub mod sanitize_maker_orders {
             &filler_key,
             0,
             oracle_price,
-            100 * PRICE_PRECISION_U64,
             clock.unix_timestamp,
             clock.slot,
         )