+++ programs/drift/src/math/matching.rs
@@ -17,15 +17,25 @@ pub fn is_maker_for_taker(
     taker_order: &Order,
     slot: u64,
 ) -> DriftResult<bool> {
+    // Maker and taker order not allowed to match if both were placed in the current slot
+    if slot == maker_order.slot && slot == taker_order.slot {
+        return Ok(false);
+    };
+
     // taker cant be post only and maker must be resting limit order
     if taker_order.post_only || !maker_order.is_resting_limit_order(slot)? {
         Ok(false)
-    // can make if taker order isn't resting (market order or limit going through auction) or if maker is post only
-    } else if !taker_order.is_resting_limit_order(slot)? || maker_order.post_only {
+    // can make if taker order isn't resting (market order or limit going through auction)
+    } else if !taker_order.is_resting_limit_order(slot)? {
         Ok(true)
     // otherwise the maker must be older than the taker order
     } else {
-        Ok(maker_order.slot < taker_order.slot)
+        Ok(maker_order
+            .slot
+            .safe_add(maker_order.auction_duration.cast()?)?
+            <= taker_order
+                .slot
+                .safe_add(taker_order.auction_duration.cast()?)?)
     }
 }