+++ programs/drift/src/state/oracle.rs
@@ -8,6 +8,9 @@ use crate::math::safe_math::SafeMath;
 use crate::math::safe_unwrap::SafeUnwrap;
 use switchboard_v2::decimal::SwitchboardDecimal;
 
+#[cfg(test)]
+mod tests;
+
 #[derive(Default, AnchorSerialize, AnchorDeserialize, Clone, Copy, Eq, PartialEq, Debug)]
 pub struct HistoricalOracleData {
     pub last_oracle_price: i64,
@@ -89,6 +92,7 @@ impl HistoricalIndexData {
 #[derive(AnchorSerialize, AnchorDeserialize, Clone, Copy, Eq, PartialEq, Debug)]
 pub enum OracleSource {
     Pyth,
+    Pyth1000,
     Switchboard,
     QuoteAsset,
 }
@@ -125,7 +129,8 @@ pub fn get_oracle_price(
     clock_slot: u64,
 ) -> DriftResult<OraclePriceData> {
     match oracle_source {
-        OracleSource::Pyth => get_pyth_price(price_oracle, clock_slot),
+        OracleSource::Pyth => get_pyth_price(price_oracle, clock_slot, 1),
+        OracleSource::Pyth1000 => get_pyth_price(price_oracle, clock_slot, 1000),
         OracleSource::Switchboard => {
             msg!("Switchboard oracle not yet supported");
             Err(crate::error::ErrorCode::InvalidOracle)
@@ -139,7 +144,11 @@ pub fn get_oracle_price(
     }
 }
 
-pub fn get_pyth_price(price_oracle: &AccountInfo, clock_slot: u64) -> DriftResult<OraclePriceData> {
+pub fn get_pyth_price(
+    price_oracle: &AccountInfo,
+    clock_slot: u64,
+    multiple: u128,
+) -> DriftResult<OraclePriceData> {
     let pyth_price_data = price_oracle
         .try_borrow_data()
         .or(Err(crate::error::ErrorCode::UnableToLoadOracle))?;
@@ -149,6 +158,13 @@ pub fn get_pyth_price(price_oracle: &AccountInfo, clock_slot: u64) -> DriftResul
     let oracle_conf = price_data.agg.conf;
 
     let oracle_precision = 10_u128.pow(price_data.expo.unsigned_abs());
+
+    if oracle_precision <= multiple {
+        msg!("Multiple larger than oracle precision");
+        return Err(crate::error::ErrorCode::InvalidOracle);
+    }
+
+    let oracle_precision = oracle_precision.safe_div(multiple)?;
 
     let mut oracle_scale_mult = 1;
     let mut oracle_scale_div = 1;