+++ tests/placeAndMakeSpotOrder.ts
@@ -156,7 +156,10 @@ describe('place and make spot order', () => {
 			marketIndex,
 			direction: PositionDirection.LONG,
 			baseAssetAmount,
-			price: new BN(40).mul(PRICE_PRECISION),
+			price: new BN(41).mul(PRICE_PRECISION),
+			auctionStartPrice: new BN(40).mul(PRICE_PRECISION),
+			auctionEndPrice: new BN(41).mul(PRICE_PRECISION),
+			auctionDuration: 10,
 			userOrderId: 1,
 			postOnly: PostOnlyParams.NONE,
 		});